backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2.16
django-cors-headers==4.8.0
django-environ==0.12.0
djangorestframework==3.16.1
dnspython==2.8.0
idna==3.10
mongoengine==0.29.1
pymongo==4.15.0
python-decouple==3.8
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
